% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_boxplot2d.R
\docType{data}
\name{GeomBoxplot2d}
\alias{GeomBoxplot2d}
\title{Create a ggplot2 geom for two dimensional box and whisker plots.}
\description{
Create a ggplot2 geom for two dimensional box and whisker plots.
}
\keyword{datasets}
