% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_error2d.R
\docType{data}
\name{GeomError2d}
\alias{GeomError2d}
\title{Create a ggplot2 geom for two dimensional error boxes.}
\description{
Create a ggplot2 geom for two dimensional error boxes.
}
\keyword{datasets}
