% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggname.R
\name{ggname}
\alias{ggname}
\title{Internal function copied from ggplot2 (not exported there).}
\usage{
ggname(prefix, grob)
}
\arguments{
\item{prefix}{The prefix part of the name.}

\item{grob}{A grob object or NULL.}
}
\description{
This function generates a unique (within-session) name for a grob, based on
the grob's class.
}
